ProgrammingClubChallenges/DiscordBot/requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
aiohttp==3.7.4.post0
requests==2.31.0
